Move HostConfig into submodule to prevent circular imports
